slideasset: Fix a bug that prevented removing assets without a thumbnail
- Fix a bug that prevented removing assets without a thumbnail. This
  bug was due to calling unlink() for the thumbnail path even when
  the path variable was NULL. This in turn caused an internal server
  error and prevented removing the slide asset.

+++ src/common/php/slide/slideasset.php
@@ -98,11 +98,19 @@ class SlideAsset extends Exportable {
 	}
 
 	public function remove() {
-		if (!empty($this->fullpath)) {
+		if (
+			!empty($this->fullpath)
+			&& is_file($this->fullpath)
+		) {
 			unlink($this->fullpath);
-			unlink($this->thumbpath);
-			$this->reset();
 		}
+		if (
+			!empty($this->thumbpath)
+			&& is_file($this->thumbpath)
+		) {
+			unlink($this->thumbpath);
+		}
+		$this->reset();
 	}
 
 	private function reset() {
